requirements.txt
python-gitlab>=3.0.0
requests>=2.25.1
configparser>=5.0.0